Refactor test and util packages
