+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -12,7 +12,7 @@
 //import org.junit.jupiter.api.io.TempDir;
 //
 //import seedu.address.logic.commands.CommandResult;
-//import seedu.address.logic.commands.ViewAllStudentsCommand;
+//import seedu.address.logic.commands.view.ViewAllStudentsCommand;
 //import seedu.address.logic.commands.exceptions.CommandException;
 //import seedu.address.logic.parser.exceptions.ParseException;
 //import seedu.address.model.Model;
