+++ src/test/java/seedu/address/logic/commands/EditLoginCommandTest.java
@@ -10,7 +10,8 @@ import static seedu.address.testutil.TypicalStudents.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.EditLoginCommand.EditLoginDescriptor;
+import seedu.address.logic.commands.edit.EditLoginCommand;
+import seedu.address.logic.commands.edit.EditLoginCommand.EditLoginDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
