+++ src/test/java/seedu/address/logic/parser/ViewCommandParserTest.java
@@ -7,10 +7,10 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailur
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.ViewAllStudentsCommand;
-import seedu.address.logic.commands.ViewCommand;
-import seedu.address.logic.commands.ViewMissionDeadlineCommand;
-import seedu.address.logic.commands.ViewOneStudentCommand;
+import seedu.address.logic.commands.view.ViewAllStudentsCommand;
+import seedu.address.logic.commands.view.ViewCommand;
+import seedu.address.logic.commands.view.ViewMissionDeadlineCommand;
+import seedu.address.logic.commands.view.ViewOneStudentCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.flag.Flag;
 import seedu.address.model.student.Name;
