+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import seedu.address.model.consultation.Consultation;
+import seedu.address.model.consultation.MasteryCheck;
 import seedu.address.model.mission.Mission;
 import seedu.address.model.quest.Quest;
 import seedu.address.model.student.Student;
@@ -121,6 +122,11 @@ public class AddressBookTest {
 
         @Override
         public ObservableList<Consultation> getConsultationList() {
+            return null;
+        }
+
+        @Override
+        public ObservableList<MasteryCheck> getMasteryChecksList() {
             return null;
         }
     }
